+++ projects/vedana/CHANGELOG.md
@@ -9,6 +9,7 @@
 * (vedana-etl) covered get_grist_data / filter_grist_data with tests
 * (vedana-etl) more fixes in processing "reference" type columns for both links
   and attrs
+* (vedana-etl) mv data filtering logic into get_grist_data, rm separate filter steps
 
 # 2025.08.29-dev.19